[LI-HOTFIX] back-port the fix for KAFKA-8950 - KafkaConsumer stops fetching
   TICKET=KAFKA-8950

   EXIT_CRITERIA = when code is rebased on top of 2.4+ to get the upstream fix

   below is the original commit from commit c8676c996ebc4a583c1efade16393ca9a73f16e5:
KAKFA-8950: Fix KafkaConsumer Fetcher breaking on concurrent disconnect (#7511)

The KafkaConsumer Fetcher can sometimes get into an invalid state where it believes that there are ongoing fetch requests, but in fact there are none. This may be caused by the heartbeat thread concurrently handling a disconnection event just after the fetcher thread submits a request which would cause the Fetcher to enter an invalid state where it believes it has ongoing requests to the disconnected node but in fact it does not. This is due to a thread safety issue in the Fetcher where it was possible for the ordering of the modifications to the nodesWithPendingFetchRequests to be incorrect - the Fetcher was adding it after the listener had already been invoked, which would mean that pending node never gets removed again.

This PR addresses that thread safety issue by ensuring that the pending node is added to the nodesWithPendingFetchRequests before the listener is added to the future, ensuring the finally block is called after the node is added.

Reviewers: Tom Lee, Jason Gustafson <jason@confluent.io>, Rajini Sivaram <rajinisivaram@googlemail.com>